requirements.txt
SpeechRecognition--3.8.1
PyAudio--0.2.11
vosk--0.3.32